+++ components/Modal/CreateUser/CreateUser.tsx
@@ -1,5 +1,6 @@
-// import classNames from 'classnames/bind';
 import { useState } from 'react';
+
+// import classNames from 'classnames/bind';
 
 import Image from 'next/image';
